pyproject.toml
[tool.poetry]
name = "rainscribe"
version = "0.1.0"
description = "Scalable automated transcription solution for Russian TV channel"
authors = ["Ark Deliev <ard@tvrain.tv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
websockets = "^12.0"
ffmpeg-python = "^0.2.0"
typing-extensions = "^4.9.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
